Moved creation of escalated case types from progress case to create case, to remove duplication and to simplify dependencies. This should make it easier to create tests for stage 2 cases.
