+++ serenity.iml
@@ -17,6 +17,12 @@
     <orderEntry type="library" scope="TEST" name="Maven: net.serenity-bdd:serenity-core:2.3.31" level="project" />
     <orderEntry type="library" scope="TEST" name="Maven: net.serenity-bdd:serenity-model:2.3.31" level="project" />
     <orderEntry type="library" scope="TEST" name="Maven: commons-net:commons-net:3.6" level="project" />
+    <orderEntry type="library" name="Maven: com.google.guava:guava:30.1-jre" level="project" />
+    <orderEntry type="library" name="Maven: com.google.guava:failureaccess:1.0.1" level="project" />
+    <orderEntry type="library" name="Maven: com.google.guava:listenablefuture:9999.0-empty-to-avoid-conflict-with-guava" level="project" />
+    <orderEntry type="library" name="Maven: org.checkerframework:checker-qual:3.5.0" level="project" />
+    <orderEntry type="library" name="Maven: com.google.errorprone:error_prone_annotations:2.3.4" level="project" />
+    <orderEntry type="library" name="Maven: com.google.j2objc:j2objc-annotations:1.3" level="project" />
     <orderEntry type="library" scope="TEST" name="Maven: com.google.inject:guice:4.2.3" level="project" />
     <orderEntry type="library" scope="TEST" name="Maven: javax.inject:javax.inject:1" level="project" />
     <orderEntry type="library" scope="TEST" name="Maven: aopalliance:aopalliance:1.0" level="project" />
@@ -184,6 +190,7 @@
     <orderEntry type="library" name="Maven: io.rest-assured:rest-assured-common:3.0.3" level="project" />
     <orderEntry type="library" name="Maven: io.rest-assured:json-schema-validator:3.0.2" level="project" />
     <orderEntry type="library" name="Maven: com.github.fge:json-schema-validator:2.2.6" level="project" />
+    <orderEntry type="library" name="Maven: com.google.code.findbugs:jsr305:3.0.0" level="project" />
     <orderEntry type="library" name="Maven: com.googlecode.libphonenumber:libphonenumber:6.2" level="project" />
     <orderEntry type="library" name="Maven: com.github.fge:json-schema-core:1.2.5" level="project" />
     <orderEntry type="library" name="Maven: com.github.fge:uri-template:0.9" level="project" />
@@ -245,10 +252,5 @@
     <orderEntry type="library" scope="RUNTIME" name="Maven: com.formdev:flatlaf:0.45" level="project" />
     <orderEntry type="library" scope="RUNTIME" name="Maven: commons-jxpath:commons-jxpath:1.3" level="project" />
     <orderEntry type="library" scope="RUNTIME" name="Maven: com.io7m.xom:xom:1.2.10" level="project" />
-    <orderEntry type="library" scope="TEST" name="Maven: com.google.guava:guava:23.0" level="project" />
-    <orderEntry type="library" name="Maven: com.google.code.findbugs:jsr305:1.3.9" level="project" />
-    <orderEntry type="library" scope="TEST" name="Maven: com.google.errorprone:error_prone_annotations:2.0.18" level="project" />
-    <orderEntry type="library" scope="TEST" name="Maven: com.google.j2objc:j2objc-annotations:1.1" level="project" />
-    <orderEntry type="library" scope="TEST" name="Maven: org.codehaus.mojo:animal-sniffer-annotations:1.14" level="project" />
   </component>
 </module>
