+++ src/test/resources/features/complaints/ComplaintsWorkstacks.feature
@@ -90,15 +90,18 @@ Feature: Complaints Workstacks
   # HOCS-3076 HOCS-3161 HOCS-4006
   @ComplaintsRegression2
   Scenario Outline: Complaints user is able to see a yellow highlighted deadline on a complaint case that is close to its deadline date
-    Given I am logged into "CS" as user "<caseType>_USER"
+    Given I am logged into "CS" as user "<user>_USER"
     When I create a single "<caseType>" case with the correspondence received date set <amountOfDays> workdays ago
     And I click to view the case in the "<workstack>" workstack
     Then the case deadline should be highlighted "yellow"
     Examples:
-      | caseType | amountOfDays | workstack              |
-      | COMP     | 15           | Complaint Registration |
-      | IEDET    | 15           | IE Detention           |
-      | SMC      | 55           | Serious Misconduct     |
+      | caseType | user | amountOfDays | workstack              |
+#      | COMP     | | 15           | Complaint Registration |
+      | COMP2      | COMP   | 15           |  Stage 2 Complaint Registration   |
+#      | IEDET    | | 15           | IE Detention           |
+#      | SMC      | |55           | Serious Misconduct     |
+#      | BF      | |15           | Border Force     |
+      | BF2      | BF      |15           | Border Force (Stage 2)    |
 
   # HOCS-3076 HOCS-3161 HOCS-4006
   @ComplaintsRegression2
